ENH: add WIP setDac, readDac and readDacExp procs
Adds procs to set a DAC, read a DAC and a debugging proc to get the
expected return value from a call to readDac.

Note: these versions are currently WIP. the `write` argument is not
yet implemented and readDacExp proc (probably) needs its output
reversed, since the Tpx3 outputs the data in LSB first order.