ENH: add wip `decode` proc to convert 2 * 32bit -> ordered 5 bytes
Adds a proc to perform the decoding of the 2 * 32bit data packets, as
they are send from the FPGA, which is an encoding for the 48bit
words the Timepix3 sends to the FPGA. However, the data is returned in
wrong order. This proc reorders them.

Currently maybe not the nicest implementation, but to see what's
happening nice to have.

TODO: this proc should later perform proper decoding into the
different components of the data, i.e. extract the command that was
executed, split data, depending on the command extract the correct information!