ENH: Add conversion procs for pixel positions
Adds three procs for converting a x/y position of a pixel into the 16 bit
address of this pixel onto the Timepix3 and vise versa the address into the x
or y position.

The proc for xy to address conversion is `xy_pixel_to_timepix3_pixel_address`
and the procs for address to x/y conversions are
`timepix3_pixel_address_to_x_pixel` and `timepix3_pixel_address_to_y_pixel`

Not sure if the seperation into two procs for conversion into x/y is the best
idea of if they schould be combined.