ENH: add `produce_columnMask` and `write_pcr` procs
The `produce_columnMask` proc gets an input of a list of columns for which the
pcr should be set. It returns the byte list for the column mask with zeros for
all selected columns and ones for all other columns (see manual v1.9 p.44).

The `write_pcr` proc creates the byte list needed for setting the PCRs on the
Timepix3 and uses the `write` proc to send it to the Timepix3. By default its
done for all columns but the columns can be selected with a list. The usage of
the `write` proc can also be turned of such that the byte list is only returned.

TODO: Currently the matrices are read pixelwise to create the byte list which is
slow. So there should be an improvement of this.