ENH: Add GeneralConfiguration procs
Adds two new procs for writing values to the GeneralConfig registers
(write_generalConfiguration) and to read them back (read_generalConfiguration).
For writing the GeneralConfig registers first the values are read from a yaml
file (which was added as `GeneralConfiguration.yml`).

Reading the yaml file could be improved. Particularly if there will be more yaml
configuration files it could be seperated in a new class.

TODO: Make it possible to set values for configuration internally such that
future functions e.g. for test pulses can set the registers as needed.

TODO: Currently the GlobalSyncHeader is always used so make it possible to
read and write the GeneralConfig registers of a selected chip.