ENH: Add `write_ctpr` proc
Adds a proc to write the column test pulse register (see manual v1.9 p.50).

TODO: Allow a selection of a chip via the Chip-ID instead of using always the
global sync header.

TODO: Because of missing documentation its unclear if columns are masked with 0
or 1. This needs a check and if required a change.