FIX: replace `columns` check in procs to include elements of list
Previously the check for the `columns` list as the argument to
`produce_column_mask`, `write_pcr` and `write_ctpr` only checked for
the length of the given list, but not for the entries of it. A call to
the function with e.g. np.arange(200, 300) was valid, but would have
failed with an `IndexError` upon accessing the BitLogic elements.