ENH: change `decode` to `decode_fpga` and add skeleton `decode` proc
This commit changes the `decode` function from the
previous (specifically written for periphery commands!) implementation
to a more general implementation, which simply performs a decoding of
a list of pairs of 32 bit words from the FPGA decoding to the native
48 bit Timepix3 decoding. It then returns a list of 6 bytes lists
corresponding to those 48 bits.

In addition a skeleton `decode` procedure is provided, which is used
to case on the header of the data returned by `decode_fpga`. Depending
on the header we know which kind of command is contained in the 48 bit
data returned by the Timepix3.

TODO: implement a dictionary, which contains the structure of the data
for each kind of command (and subcommand, which itself depends on the
Timepix3 general settings in use)