ENH: add attributes for general config to Tpx3 and load during `init`
Now we load the YAML general configuration file already during
initialization of the Timepix3 class. The content is stored as a
nested dictionary in the class and also each value is stored
individually as an attribute for faster access.

Also adds a proc `reset_config_attributes` which simply sets all of
these attributes to the default value the Tpx3 has after reset.

A single `get_configuration_register` proc is also added to allow the
creation of the 12 bit BitLogic word based on the read configuration file.