ENH: Add a proc for decoding 48 bit words
Changes the proc 'decode' such that it decodes a 48 bit word based on the header
of the word and the formation of the words for different commands (see manual
v1.9 p.28). The proc gets the 48 bit data composed by a list of 6 bytes and
additionally the expected header as binary, hex or int. If the expected and
the received header match the word is decoded if not an error is raised.
The proc returns a list of bitarrays depending on the header.

The following lists are used:

                        dataout[0]      dataout[1]      dataout[2]      dataout[3]
- Acquisition*:         Address[15:0]   TOA[13:0]       TOT[9:0]        FTOA[3:0]
- @ data readout:       Dummy[43:0]         -               -               -
- CTPR configuration:   Address[6:0]    EoC[17:0]       CTPR[1:0]           -
- Pixel configuration:  Address[15:0]   Config[5:0]         -               -
- Periphery command:    DataOut[39:0]       -               -               -
- Control command:      H1,H2,H3[7:0]   ChipID[31:0]        -               -

* Note that for acquisition the meaning of dataout[1] to dataout[3] depends on the
  settings of the general config registers 'Op_mode' and 'Fast_Io_en' (see manual
  v1.9 p.40). The composition of the list is independent of this, only the
  interpretation must change accordingly.

TODO: This implementation is not the nicest. One can change from a hardcoding
      for the different commands to a dictionarry. Also it would be nicer of the
      'decode_fpga' returns a bitarray such that no conversion to a bitarray is
      needed in the 'decode' proc