FIX: Limit range of thresholds in threshold_scan
For low thresholds 32 bit words are missed in the FIFO. A workaround for now
is to limit the threshold scan to a working threshold range. A sleep before
reading the FIFO extends the threshold range a bit but sleeps up to one second
do not solve the problem for the whole range.