ENH: add `testChipId.py` to try to read the chips ChipID
Explanation for the determination of this chips ChipID:
The idea is to send a valid command using a local chip header with a
chip ID different from the one the chip has. According to the manual
page 29 of manual v1.9, a chip will send the `OtherChipCommand` containing
its own ChipID. So therefore we send the SenseDACsel command, which with
a local chip ID of 0x01. The chip correctly answers with an OtherChipCommand,
but containing the chip ID with all 0s...