MAINT: replace explicit set of _config by loop, change reset attrs.
We replace the explicit setting of the _config dictionary of the
contents of the YAML file, by a loop over the self.config dictionary,
which simply contains the data from the YAML file.

The `reset_config_attributes` proc is changed to simply reset these
values to the either default values (to_default == True) of the YAML
file or the user specified settings (mode values, to_default == False).