ENH: Add the outputBlock configuration
Adds a dictionary for the outputBlock configuration based on the CustomDict
which is already used for the DACs and the generalConfig. Furthermore the procs
'read_outputBlock_yaml' for reading a corresponding yaml file and
'write_outputBlock_config' for writing the internal condiguration to the chip
were added.

Adds also the needed yaml file for the outputBlock configuration.