ENH: unify all `read_*_yaml` procs into one generic solution
As a requirement for this to work, the identifiers of all registers in
the YAML files were set to the same value, i.e. DACs are not also
referred by 'address', instead of 'code'. And all use 'value' instead
of 'mode' now.

This is done by a generic function `read_yaml`, which receives the
dictionary type to be created from a given YAML file. Given a
`dictName` dictionary, which stores the names of the attributes, which
are going to be used, the YAML file is parsed, dictionaries added and
assigend.