ENH: Add op_mode and VCO_mode to h5 files of scan_base
For iterpreting the raw data the mode of the chip is needed. Its a combination
of the two general config parameters Op_mode (TOT & TOA, only TOA, event counter)
and Fast_Io_en (enable, disable). Such that the mode can be used for raw data
interpretation it is now stored in the h5 file.