ENH: Add mode dependent data interpretation in analysis.py
Changes the '_interpret_raw_data' in analysis.py such that it interprets the
data dependent on the mode which is defined by 'mode' and 'vco_mode'.
Therefore the mode is read from the h5 file in scan_threshold.py and then
given to 'interpret_raw_data'.
Furthermore '_interpret_raw_data' changes such that it can be used with njit.
Therefore the parameter 'pix_data' was addded which contains names ad data
formates for the different pixel data.