[ENH] refactor scan_base.start and take_data.scan to auto reset chip
Auto resets the chip if 255 or more RX errors were received.

Refactors the scan_base.start and take_data.scan procedure by
splitting out initialization and stop parts.

Scan calls a `scan_impl`, which performs the actual scan, but returns
`False`, if the chip needs to be reset. In that case `scan` handles
the reset and the proc is called again.