ENH: fist iteration of multichip support for lohengrin
With this commit the general communication is shifted from global headers
to local headers and the configuration structure is adapted for multiple
chips. So far, the GUI and within it the hardware init, all settings and
data taking are supported. Other scans and the event display are not
supported yet. The CLI was not tested so far.
All tests were performed with only a single chip but communication
via local headers. Functionality with multiple chips still has to be
tested.

Co-authored-by: Thomas Block <thomas.matt.block@gmail.com>